[fix] 글로벌 스케일 변경 함수가 변동 값을 입력받도록 함

+++ src/main/kotlin/composable/intro/Intro.kt
@@ -43,7 +43,7 @@ fun Intro(
     localApplicationData: LocalDataState,
     openRecent: (DoodlerEditorType, File) -> Unit,
     openSelector: (DoodlerEditorType) -> Unit,
-    changeGlobalScale: () -> Unit
+    changeGlobalScale: (Float) -> Unit
 ) {
 
     var settingsVisible by remember { mutableStateOf(false) }
